Keep node-map layout helpers out of the PixiJS import graph
The new nodeLayout test failed in CI with "navigator is not defined": it
imported nodeCenter/snapToTile from NodeMapRederer.tsx, which imports
pixi.js at module scope, and vitest's first project runs *.test.ts under
environment: 'node'. It passed locally only because Node 21+ provides a
global navigator that CI's Node does not.

Move the pure grid geometry — the layout constants, snapToTile,
nodeCenter and startPos — into nodeLayout.ts, which depends only on
TILE_SIZE and pulls in no PixiJS. The renderer imports them from there,
and the test no longer reaches a browser-only module at all.

Also guard app.ticker?.add in buildNodeMarker. The sprite load above it
is awaited, so the Pixi app can be destroyed before the tick is
registered, leaving ticker null — a pre-existing teardown race that
surfaced as an unhandled rejection in the same CI run.
